backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
anthropic>=0.42.0
Pillow>=10.0.0
twilio>=9.0.0
resend>=2.0.0
gunicorn>=23.0.0
